Fix payout error handler

+++ src/Facades/Traits/PayUPayouts.php
@@ -2,7 +2,6 @@
 
 namespace xGrz\PayU\Facades\Traits;
 
-use xGrz\PayU\Api\Exceptions\PayUGeneralException;
 use xGrz\PayU\Enums\PayoutStatus;
 use xGrz\PayU\Facades\Config;
 use xGrz\PayU\Jobs\SendPayoutJob;
@@ -11,22 +10,39 @@ use xGrz\PayU\Models\Payout;
 
 trait PayUPayouts
 {
+    /**
+     * Store payout data in database.
+     * Dispatches PayoutCreated event
+     *
+     * @param int|float $amount
+     * @return bool
+     */
     public static function payout(int|float $amount): bool
     {
         if (!Config::getShopId()) return false;
 
         try {
-            return (bool)Payout::create(['amount' => $amount]);
+            return (bool) Payout::create(['amount' => $amount]);
         } catch (\Exception $e) {
             return false;
         }
     }
 
+    public static function payoutStatusCheck(Payout $payout, int $delay = null): bool
+    {
+        if (!Config::getShopId()) return false;
+        if (!$payout->status->hasAction('processing')) return false;
+
+        UpdatePayoutStatusJob::dispatch($payout)
+            ->delay(is_null($delay) ? Config::getPayoutInterval() : $delay);
+
+        return true;
+    }
+
     public static function retryPayout(Payout $payout, int $delay = null): bool
     {
-        if (!$payout->status->hasAction('retry')) {
-            throw new PayUGeneralException('Cannot retry payout. Action [retry] impossible');
-        }
+        if (!Config::getShopId()) return false;
+        if (!$payout->status->hasAction('retry')) return false;
 
         $payout
             ->update(['status' => PayoutStatus::RETRY]);
@@ -38,15 +54,9 @@ trait PayUPayouts
         return true;
     }
 
-    public static function payoutStatusCheck(Payout $payout, int $delay = null): bool
-    {
-        UpdatePayoutStatusJob::dispatch($payout)
-            ->delay(is_null($delay) ? Config::getPayoutInterval() : $delay);
-
-        return true;
-    }
     public static function cancelPayout(Payout $payout): bool
     {
+        if (!Config::getShopId()) return false;
         if (!$payout->status->hasAction('delete')) return false;
 
         $payout->delete();
